Spin off generators to different script
